refactor: separate PyO3 from pure Rust core
- Move all PyO3 dependencies to src/python_bridge.rs
- Remove PyO3 imports from core modules (lib.rs, error.rs, commitment.rs, scoring.rs, round.rs)
- Enable clean compilation with --no-default-features (pure Rust)
- Maintain full Python API compatibility via python_bridge module
- Add convenience method get_image_embedding() to ScoreValidator
- Update Cargo.toml with better feature documentation
- Add comprehensive architecture guidelines to RUST_README.md
- All 28 tests pass in both pure Rust and Python modes

Benefits:
- Faster development cycle (no PyO3 compilation overhead)
- Better testing (pure Rust logic independent of Python)
- Easier to add other language bindings (C FFI, WASM, etc)
- Cleaner architecture following SOLID principles
- Clear development workflow for future contributors
